If the data folder doesn't exist, don't try to expand files from it.

+++ TDoR/views/pages/rebuild.php
@@ -157,40 +157,45 @@
         echo('Adding dummy data...<br>');
 
         // Prescan - look for zip files and extract them
-        $filenames = scandir('data');
-
-        foreach ($filenames as $filename)
+        $data_folder = 'data';
+        
+        if (file_exists($data_folder) )
         {
-            $fileext = pathinfo($filename, PATHINFO_EXTENSION);
+            $filenames = scandir($data_folder);
 
-            if (0 == strcasecmp('zip', $fileext) )
+            foreach ($filenames as $filename)
             {
-                extract_zipfile('data/'.$filename);
+                $fileext = pathinfo($filename, PATHINFO_EXTENSION);
+
+                if (0 == strcasecmp('zip', $fileext) )
+                {
+                    extract_zipfile('data/'.$filename);
+                }
             }
+
+            // Now look for csv files and import them
+            $filenames = scandir($data_folder);
+
+            echo count($filenames).' files found in data folder<br>';
+
+            foreach ($filenames as $filename)
+            {
+                $fileext = pathinfo($filename, PATHINFO_EXTENSION);
+
+                if (0 == strcasecmp('csv', $fileext) )
+                {
+                    echo("Importing data from $filename...<br>");
+
+                    add_data_from_file($db, 'data/'.$filename);
+                }
+                else
+                {
+                    echo("Skipping $filename<br>");
+                }
+            }
+
+            create_homepage_slider_images();
         }
-
-        // Now look for csv files and import them
-        $filenames = scandir('data');
-
-        echo count($filenames).' files found in data folder<br>';
-
-        foreach ($filenames as $filename)
-        {
-            $fileext = pathinfo($filename, PATHINFO_EXTENSION);
-
-            if (0 == strcasecmp('csv', $fileext) )
-            {
-                echo("Importing data from $filename...<br>");
-
-                add_data_from_file($db, 'data/'.$filename);
-            }
-            else
-            {
-                echo("Skipping $filename<br>");
-            }
-        }
-
-        create_homepage_slider_images();
     }
 ?>
